Some hackery to fix the javadoc build for hdfs-proxy, which could not be deployed to maven central with the rest of 1.9.14 because the JDK's module name derivation algorithm fails on hadoop-shaded-protobuf_3_7.
This commit bumps some dependencies to remove one problem child (an ancient incubating version of hadoop-htrace) but brings
us this new problem, fixed by setting the javadoc source level to JDK 8.

I have no idea if there are tests for this thing somewhere, so not deploying it to Maven central, but it at least
builds correctly and the javadoc build works, so it _could_ be deployed to Maven central.