Remove hdfs-proxy and friends from the build
Due to the various pathologies in dependencies of hadoop-common, including

  * dnsjava putting classes in the default package
  * hacking ServiceLoader to get it to load classes that have constructor arguments
  * use of non-public `sun.*` classes that do not exist past JDK 8

it is likely impossible for this project ever to be modularized without a number
of its dependencies undergoing significant rewrites.

It is impossible - particularly when building javadoc (which is not even needed
for this project that is a standalone application, not public API - but javadoc
is required for publishing on Maven central) - for non-modular and modular
projects to coexist in the same tree / build.

As such, this project needs to be moved to its own git repository and not
be included here.

c.f. https://github.com/Telenav/kivakit/issues/125#issuecomment-1189552612